Move all Elastic Charts docs from the `Component` tab to the `Data visualization` tab
+ remove unnecessary badge with charts version (handle that holistically if we're going to add it back in)

+ sidebar and copy tweaks
