--- packages/website/docs/components/elastic_charts/_category_.yml
@@ -1,3 +1,0 @@
-label: Elastic Charts
-position: 10
-collapsed: false
