+++ packages/website/docs/data_visualization/accessibility_features.mdx
@@ -1,6 +1,7 @@
 ---
-slug: /elastic-charts/accessibility-features
+slug: /charts/accessibility
 id: elastic_charts_accessibility_features
+sidebar_position: 3
 ---
 
 # Accessibility features
@@ -9,8 +10,6 @@ id: elastic_charts_accessibility_features
 import * as ElasticCharts from '@elastic/charts';
 import { useChartBaseTheme } from './use_chart_base_theme.ts';
 ```
-
-<Badge href="https://github.com/elastic/elastic-charts">External library: @elastic/charts v64.1.0</Badge>
 
 [Elastic Charts](https://elastic.github.io/elastic-charts) provides some accessibilty features for all users.
