+++ packages/website/docs/data_visualization/dashboard_good_practices.mdx
@@ -1,6 +1,7 @@
 ---
 id: dashboard_good_practices
-slug: /data-viz/dashboard-good-practices
+slug: /charts/dashboard-good-practices
+sidebar_position: 2
 ---
 
 # Dashboard good practices
