+++ packages/website/docs/data_visualization/elastic_charts_design_system.mdx
@@ -1,6 +1,7 @@
 ---
 id: elastic_charts_design_system
-slug: /data-viz/elastic-charts-design-system
+slug: /charts/design-system
+sidebar_position: 1
 ---
 
 # Elastic Charts as a design system
