+++ packages/website/docs/data_visualization/sizing.mdx
@@ -1,11 +1,10 @@
 ---
 slug: /elastic-charts/sizing
 id: elastic_charts_sizing
+sidebar_position: 5
 ---
 
 # Sizing
-
-<Badge>External library: @elastic/charts v64.1.0</Badge>
 
 When placing charts into smaller containers or panels, you must re-evaluate your data to provide a more simplified version. This could be as simple as shifting legend positions from the right side to the bottom or adding annotations to give context and describe the key points in your data.
