+++ packages/website/docs/data_visualization/types/_category_.yml
@@ -1,0 +1,3 @@
+label: Types
+collapsed: false
+position: 4
