+++ packages/website/docs/data_visualization/types/categorical.mdx
@@ -1,11 +1,10 @@
 ---
-slug: /elastic-charts/categorical
+slug: /charts/types/categorical
 id: elastic_charts_categorical
+sidebar_position: 1
 ---
 
 # Categorical
-
-<Badge>External library: @elastic/charts v64.1.0</Badge>
 
 Categorical charts compare data between multiple distinct categories. Avoid using a line chart because it might create confusion with a time series chart. Use a bar chart instead.
