+++ packages/website/docs/data_visualization/types/metric_anatomy.png
