+++ packages/website/docs/data_visualization/types/metric_chart.mdx
@@ -1,17 +1,16 @@
 ---
-slug: /elastic-charts/metric-chart
+slug: /charts/types/metric-chart
 id: elastic_charts_metric_chart
+sidebar_position: 1
 ---
 
 # Metric Chart
 
 ```mdx-code-block
-import { useChartBaseTheme } from './use_chart_base_theme';
+import { useChartBaseTheme } from '../use_chart_base_theme';
 import * as ElasticCharts from '@elastic/charts';
 import BrowserOnly from '@docusaurus/BrowserOnly';
 ```
-
-<Badge>External library: @elastic/charts v64.1.0</Badge>
 
 The **Metric** chart is a single value visualization available from [Elastic Charts](https://elastic.github.io/elastic-charts). It allows you to represent a metric, a KPI, a specific state with a very simple interface, similar to [**EuiStat**](/docs/display/stat) but with more data visualization oriented capabilities.
