+++ packages/website/docs/data_visualization/types/metric_intro.png
