+++ packages/website/docs/data_visualization/types/time_series.mdx
@@ -1,11 +1,10 @@
 ---
-slug: /elastic-charts/time-series
+slug: /charts/types/time-series
 id: elastic_charts_time_series
+sidebar_position: 1
 ---
 
 # Time series
-
-<Badge>External library: @elastic/charts v64.1.0</Badge>
 
 Time series charts show data over a period of time, such as trends or comparisons across multiple categories. When smaller changes exist, it’s better to use line charts rather than bar charts.
